+++ sdk/src/main/java/com/bugsnag/android/DefaultDelivery.java
@@ -15,6 +15,7 @@ import java.util.Map;
 
 class DefaultDelivery implements Delivery {
 
+    private static final int HTTP_REQUEST_FAILED = 0;
     @Nullable private final ConnectivityManager connectivityManager;
 
     DefaultDelivery(@Nullable ConnectivityManager connectivityManager) {
@@ -80,6 +81,9 @@ class DefaultDelivery implements Delivery {
             return conn.getResponseCode();
         } catch (IOException exception) {
             throw new DeliveryFailureException("IOException encountered in request", exception);
+        } catch (Exception exception) {
+            Logger.warn("Unexpected error delivering payload", exception);
+            return HTTP_REQUEST_FAILED;
         } finally {
             IOUtils.close(conn);
         }